crude_oil_forecast/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
tensorflow>=2.13.0
keras>=2.13.0
streamlit>=1.28.0
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0
requests>=2.31.0
python-dotenv>=1.0.0
fredapi>=0.5.1
yfinance>=0.2.28
ta>=0.10.2
joblib>=1.3.0
scipy>=1.11.0
statsmodels>=0.14.0
textblob>=0.17.1
newsapi-python>=0.2.7
